bookplots and static functions added
